Is good Working is step college

+++ WebStoreMVC/WebStoreMVC/Areas/Admin/Models/Product/CreateProductImageViewModel.cs
@@ -3,5 +3,5 @@
 public class CreateProductImageViewModel
 {
     public string Name { get; set; } = string.Empty;
-    public int Priority { get; set; }
+    public short Priority { get; set; }
 }
